Add property edgeWidthFunction to tf-graph
This change adds a `edgeWidthFunction` property to tf-graph and
tf-graph-board. This is a method that a user can specify to compute the
thickenss (in pixels) of individual edges in the graph. If specified,
edge thickness is no longer based on tensor size but rather on
`edgeWidthFunction`.

This change is part of making the graph generalizable to other
applications.

Test plan: Make `tf-graph-dashboard` pass some arbitrary
`edgeWidthFunction` to `tf-graph-board`.

Start TensorBoard with a graph that normally varies in edge thicknes.
Note that the edge thicknesses vary based on the provided function
instead of tensor size.